fix: execute pre- and post-bench scripts for non-perf walltime runner
